# Store the hashValue to avoid repeating the hash calculation during the hash resize.