Discuss:
Cache the hash value in lh_entry. This value is used directly during hash resize to calculate the index of the data in hash table. This method will reduce the times of hash calculations. The improving effect of the parse speed is more obvious in the case of multiple objects. But this method will add 4 bytes overhead for every lh_entry in 32-bit systems.

|json file|ave_time before/ms|ave_time after/ms|optimization ratio/%|
| ---- | ---- | ---- | ---- |
|[twitter.json](https://github.com/dota17/nativejson-benchmark/blob/master/data/twitter.json)|11.66|10.53|9.69|
|jeopardy_50m.json|964|920|4.5|
|[citm_catalog_1.6m.json](https://github.com/dota17/nativejson-benchmark/blob/master/data/citm_catalog.json)|28.82|27|6.3|