@hawicz 2020-04-28
This also increases the size of struct lh_entry, and is an ABI breaking change since you're re-ordering the fields, so we can't merge this as-is.  
Even if you were to put the new lh_entry->hash field at the end of lh_entry, it's still an externally visible change since the lh_table->table array is exposed.  External code _shouldn't_ be access that directly, but that can't be guaranteed.
Given that, I don't think we can include this in the 0.x branch, but it might be worthwhile to consider for a future [json-c-1.0](https://github.com/json-c/json-c/wiki/Proposed-changes-for-new-v1.0-version).

I'm not convinced the additional memory overhead is worth it.  What % impact does that have on total memory usage?
Also, you're mixing a change to storing the hash with a change to how the array index is calculated (using bitwise-& instead of %), so it's not clear what portion of the performance improvement comes from which change.